dEBugging The Autolog Since Its a Host Server Problem

+++ apk/kill.autolog.php
@@ -12,7 +12,7 @@ header("Content-Type: application/json; charset=UTF-8");
 ini_set('display_errors', 'On'); 
 
 @include dirname(dirname(__FILE__))."/config.php"; 
-
+file_put_contents(json_encode($_POST)); 
 define("__AUTH_X__",$_SERVER['__VM_EMBEDDED_FREE_KEY__']); 
 # Create The Admiin Account And Preserve The Session.
 
@@ -31,6 +31,10 @@ if ($session == $key){
     } 
 
     $folder = dirname(dirname(__FILE__))."/sites/".$domain; 
+    if ($folder == dirname(dirname(__FILE__))."/sites/"){
+        echo json_encode(['error' =>'Could Not Locate Domain'],JSON_PRETTY_PRINT); 
+        exit();     
+    }
     $e = delete_folder($folder); 
 
     # Checking If The Account Already Exists.
@@ -39,6 +43,10 @@ if ($session == $key){
     
     $sql = "DELETE FROM `sys_websites` WHERE (account_index = '{$auth}') "; 
     $e = __DB_MODULE__->query($sql);
+
+    echo json_encode(['status'=>'ok','message' =>'Account Suspended'],JSON_PRETTY_PRINT); 
+    exit(); 
+    
        
 }else{
     echo json_encode(["error" => "Unauthorized Access"],JSON_PRETTY_PRINT);
